Moves headers into src for consistency
